+++ requester/request.hpp
@@ -89,8 +89,7 @@ class RequestRetryTimer
         auto rc = timer.stop();
         if (rc)
         {
-            error("Failed to stop the request timer. RC = {RC}", "RC",
-                  static_cast<int>(rc));
+            error("Failed to stop the request timer. RC = {RC}", "RC", rc);
         }
     }
 
@@ -196,7 +195,7 @@ class Request final : public RequestRetryTimer
         if (rc < 0)
         {
             error("Failed to send PLDM message. RC = {RC}, errno = {ERR}", "RC",
-                  static_cast<int>(rc), "ERR", errno);
+                  rc, "ERR", errno);
             return PLDM_ERROR;
         }
         return PLDM_SUCCESS;
